Move handler classes to handler package
