+++ src/main/java/jtorrent/domain/handler/PeerHandler.java
@@ -1,4 +1,4 @@
-package jtorrent.domain.model.peer;
+package jtorrent.domain.handler;
 
 import static jtorrent.domain.Constants.PEER_ID;
 
@@ -9,6 +9,7 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
+import jtorrent.domain.model.peer.Peer;
 import jtorrent.domain.model.peer.exception.InfoHashMismatchException;
 import jtorrent.domain.model.peer.message.Handshake;
 import jtorrent.domain.model.peer.message.typed.Bitfield;
